+++ lib/client/lmp.dart
@@ -13,15 +13,12 @@ Future<TimeSeries<num>> getHourlyLmpCaiso(
     required LmpComponent component,
     required Term term,
     required String rustServer}) async {
-  final url = '$rustServer/caiso/prices/${market.name.toLowerCase()}/hourly/'
-      'start/${term.startDate.toString()}/end/${term.endDate.toString()}'
-      '?node_ids=$locationName&components=${component.shortName()}';
-  var response = await http.get(Uri.parse(url));
-  var data = json.decode(response.body) as List;
-  return TimeSeries.fromIterable(data.map((e) => IntervalTuple<num>(
-      Hour.beginning(
-          TZDateTime.parse(Caiso.location, e['hour_beginning'])),
-      e['price'])));
+  return Caiso().getHourlyLmp(
+      market: market,
+      locationName: locationName,
+      component: component,
+      term: term,
+      rustServer: rustServer);
 }
 
 Future<TimeSeries<num>> getHourlyLmpIeso(
@@ -30,15 +27,12 @@ Future<TimeSeries<num>> getHourlyLmpIeso(
     required LmpComponent component,
     required Term term,
     required String rustServer}) async {
-  final url = '$rustServer/ieso/prices/${market.name.toLowerCase()}/hourly/'
-      'start/${term.startDate.toString()}/end/${term.endDate.toString()}'
-      '?locations=$locationName&components=${component.shortName()}';
-  var response = await http.get(Uri.parse(url));
-  var data = json.decode(response.body) as List;
-  return TimeSeries.fromIterable(data.map((e) => IntervalTuple<num>(
-      Hour.beginning(
-          TZDateTime.parse(Ieso.location, e['hour_beginning'])),
-      e['price'])));
+  return Ieso().getHourlyLmp(
+      market: market,
+      locationName: locationName,
+      component: component,
+      term: term,
+      rustServer: rustServer);
 }
 
 Future<TimeSeries<num>> getHourlyLmpIsone(
@@ -47,15 +41,12 @@ Future<TimeSeries<num>> getHourlyLmpIsone(
     required LmpComponent component,
     required Term term,
     required String rustServer}) async {
-  final url = '$rustServer/isone/prices/${market.name.toLowerCase()}/hourly/'
-      'start/${term.startDate.toString()}/end/${term.endDate.toString()}'
-      '?ptids=$ptid&components=${component.shortName()}';
-  var response = await http.get(Uri.parse(url));
-  var data = json.decode(response.body) as List;
-  return TimeSeries.fromIterable(data.map((e) => IntervalTuple<num>(
-      Hour.beginning(
-          TZDateTime.parse(IsoNewEngland.location, e['hour_beginning'])),
-      e['price'])));
+  return IsoNewEngland().getHourlyLmp(
+      market: market,
+      ptid: ptid,
+      component: component,
+      term: term,
+      rustServer: rustServer);
 }
 
 Future<TimeSeries<num>> getHourlyLmpNyiso(
@@ -64,15 +55,12 @@ Future<TimeSeries<num>> getHourlyLmpNyiso(
     required LmpComponent component,
     required Term term,
     required String rustServer}) async {
-  final url = '$rustServer/nyiso/prices/${market.name.toLowerCase()}/hourly/'
-      'start/${term.startDate.toString()}/end/${term.endDate.toString()}'
-      '?ptids=$ptid&components=${component.shortName()}';
-  var response = await http.get(Uri.parse(url));
-  var data = json.decode(response.body) as List;
-  return TimeSeries.fromIterable(data.map((e) => IntervalTuple<num>(
-      Hour.beginning(
-          TZDateTime.parse(NewYorkIso.location, e['hour_beginning'])),
-      e['price'])));
+  return NewYorkIso().getHourlyLmp(
+      market: market,
+      ptid: ptid,
+      component: component,
+      term: term,
+      rustServer: rustServer);
 }
 
 extension CaisoLmpPriceExtension on Caiso {
@@ -91,6 +79,42 @@ extension CaisoLmpPriceExtension on Caiso {
         Hour.beginning(
             TZDateTime.parse(preferredTimeZoneLocation, e['hour_beginning'])),
         e['price'])));
+  }
+
+  Future<TimeSeries<num>> getDailyLmp(
+      {required Market market,
+      required String locationName,
+      required LmpComponent component,
+      required Term term,
+      required Bucket bucket,
+      required String rustServer}) async {
+    final url = '$rustServer/caiso/prices/${market.name.toLowerCase()}/daily/'
+        'start/${term.startDate.toString()}/end/${term.endDate.toString()}'
+        '?node_ids=$locationName&components=${component.shortName()}'
+        '&buckets=${bucket.name}';
+    var response = await http.get(Uri.parse(url));
+    var data = json.decode(response.body) as List;
+    return TimeSeries.fromIterable(data.map((e) => IntervalTuple<num>(
+        Date.fromIsoString(e['date'], location: preferredTimeZoneLocation),
+        e['value'])));
+  }
+
+  Future<TimeSeries<num>> getMonthlyLmp(
+      {required Market market,
+      required String locationName,
+      required LmpComponent component,
+      required Term term,
+      required Bucket bucket,
+      required String rustServer}) async {
+    final url = '$rustServer/caiso/prices/${market.name.toLowerCase()}/monthly/'
+        'start/${term.startDate.toString()}/end/${term.endDate.toString()}'
+        '?node_ids=$locationName&components=${component.shortName()}'
+        '&buckets=${bucket.name}';
+    var response = await http.get(Uri.parse(url));
+    var data = json.decode(response.body) as List;
+    return TimeSeries.fromIterable(data.map((e) => IntervalTuple<num>(
+        Month.fromIsoString(e['month'], location: preferredTimeZoneLocation),
+        e['value'])));
   }
 }
 
@@ -111,6 +135,22 @@ extension IesoLmpPriceExtension on Ieso {
             TZDateTime.parse(preferredTimeZoneLocation, e['hour_beginning'])),
         e['price'])));
   }
+
+  Future<TimeSeries<num>> getDailyLmp(
+      {required Market market,
+      required int ptid,
+      required LmpComponent component,
+      required Term term,
+      required Bucket bucket,
+      required String rustServer}) async {
+    final url = '$rustServer/ieso/prices/${market.name.toLowerCase()}/daily/'
+        'start/${term.startDate.toString()}/end/${term.endDate.toString()}'
+        '?ptids=$ptid&components=${component.shortName()}&buckets=${bucket.name}';
+    var response = await http.get(Uri.parse(url));
+    var data = json.decode(response.body) as List;
+    return TimeSeries.fromIterable(data.map((e) => IntervalTuple<num>(
+        Date.fromIsoString(e['date'], location: Ieso.location), e['value'])));
+  }
 }
 
 extension IsoneLmpPriceExtension on IsoNewEngland {
@@ -130,6 +170,40 @@ extension IsoneLmpPriceExtension on IsoNewEngland {
             TZDateTime.parse(preferredTimeZoneLocation, e['hour_beginning'])),
         e['price'])));
   }
+
+  Future<TimeSeries<num>> getDailyLmp(
+      {required Market market,
+      required int ptid,
+      required LmpComponent component,
+      required Term term,
+      required Bucket bucket,
+      required String rustServer}) async {
+    final url = '$rustServer/isone/prices/${market.name.toLowerCase()}/daily/'
+        'start/${term.startDate.toString()}/end/${term.endDate.toString()}'
+        '?ptids=$ptid&components=${component.shortName()}&buckets=${bucket.name}';
+    var response = await http.get(Uri.parse(url));
+    var data = json.decode(response.body) as List;
+    return TimeSeries.fromIterable(data.map((e) => IntervalTuple<num>(
+        Date.fromIsoString(e['date'], location: IsoNewEngland.location),
+        e['value'])));
+  }
+
+  Future<TimeSeries<num>> getMonthlyLmp(
+      {required Market market,
+      required int ptid,
+      required LmpComponent component,
+      required Term term,
+      required Bucket bucket,
+      required String rustServer}) async {
+    final url = '$rustServer/isone/prices/${market.name.toLowerCase()}/monthly/'
+        'start/${term.startDate.toString()}/end/${term.endDate.toString()}'
+        '?ptids=$ptid&components=${component.shortName()}&buckets=${bucket.name}';
+    var response = await http.get(Uri.parse(url));
+    var data = json.decode(response.body) as List;
+    return TimeSeries.fromIterable(data.map((e) => IntervalTuple<num>(
+        Month.fromIsoString(e['month'], location: IsoNewEngland.location),
+        e['value'])));
+  }
 }
 
 extension NyisoLmpPriceExtension on NewYorkIso {
@@ -148,5 +222,39 @@ extension NyisoLmpPriceExtension on NewYorkIso {
         Hour.beginning(
             TZDateTime.parse(preferredTimeZoneLocation, e['hour_beginning'])),
         e['price'])));
+  }
+
+  Future<TimeSeries<num>> getDailyLmp(
+      {required Market market,
+      required int ptid,
+      required LmpComponent component,
+      required Term term,
+      required Bucket bucket,
+      required String rustServer}) async {
+    final url = '$rustServer/nyiso/prices/${market.name.toLowerCase()}/daily/'
+        'start/${term.startDate.toString()}/end/${term.endDate.toString()}'
+        '?ptids=$ptid&components=${component.shortName()}&buckets=${bucket.name}';
+    var response = await http.get(Uri.parse(url));
+    var data = json.decode(response.body) as List;
+    return TimeSeries.fromIterable(data.map((e) => IntervalTuple<num>(
+        Date.fromIsoString(e['date'], location: NewYorkIso.location),
+        e['value'])));
+  }
+
+  Future<TimeSeries<num>> getMonthlyLmp(
+      {required Market market,
+      required int ptid,
+      required LmpComponent component,
+      required Term term,
+      required Bucket bucket,
+      required String rustServer}) async {
+    final url = '$rustServer/nyiso/prices/${market.name.toLowerCase()}/monthly/'
+        'start/${term.startDate.toString()}/end/${term.endDate.toString()}'
+        '?ptids=$ptid&components=${component.shortName()}&buckets=${bucket.name}';
+    var response = await http.get(Uri.parse(url));
+    var data = json.decode(response.body) as List;
+    return TimeSeries.fromIterable(data.map((e) => IntervalTuple<num>(
+        Month.fromIsoString(e['month'], location: NewYorkIso.location),
+        e['value'])));
   }
 }